[Test Infrastructure] Upgraded Test Interface and Output of Results (#66)
### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
### 1. **Improved Test Output with `pytest-sugar`**
**Visual Progress + Concise Test Summary**
Upgraded the look and readability of test output using `pytest-sugar`.
Here we ca track which test cases pass, skip or fail and on the side we
have an active overview of how our tests are progressing i.e what
percentage of tests have completed. At the end we get a clear and
concise overview of all the tests we ran and their result summary.
### 2. **Automated Test Configuration with `pytest.ini`**
**Simplified Command-Line Invocation + Increased Verbosity**
The test configuration has been automated via the `pytest.ini` file.
This eliminates the need to manually add flags (like `-v`) each time
tests are run. The `-v` verbosity flag is set up by default in the
`pytest.ini` file, ensuring that test parameters are clearly displayed
in the output, making it easier to track the setup, state and condition
of each test.
* added feature to run all tests by simply typing `pytest` in terminal
(under python_tests directory of course).
### 3. **Centralized Error Logging with `pytest_errors.log`**
 **Clean Terminal Output + Error Logging**
All error messages and warnings are now redirected to a dedicated log
file (`pytest_errors.log`). This ensures the terminal only shows a
concise summary of the test results. Removed clutter of error messages
and warnings in terminal. We are unable to see which tests have
passed/failed/skipped. Errors and failures are logged so they can be
reviewed later without cluttering the terminal output.
### 4. **Clearer Test Parameter Display**
**Full Parameter Visibility + Enhanced Debugging**
The output now clearly shows all parameters passed to the test function,
including data types, states, and conditions. This eliminates the
confusion over which parameters were used in the tests. By displaying
the parameters, we can better understand the state and conditions under
which the tests pass or fail, making it easier to debug issues.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

---------

Co-authored-by: Filip Vranic <fvranic@tenstorrent.com>