#0: revert stateful api change for worker sync (#19704)
Problem:
after using statful api, the cmd buf seems to be changing values, the
golden value for noc ctrl should be 8346, however, when running the
update ptr, the values change to different values, depending on which
machine it is running on. On t3k, it changed to 1170, and 9370, looks
like trid field is being set, and inline write field is set to 0.
on TG, it changes to 12442. 
Can't pin-point which code path is modifying the cmd buf value


### Checklist
- [x] [All post commit]
https://github.com/tenstorrent/tt-metal/actions/runs/14086885456
- [x] benchmark
https://github.com/tenstorrent/tt-metal/actions/runs/14086862922
- [x] t3k unit
https://github.com/tenstorrent/tt-metal/actions/runs/14086879132