Fix submesh tests where submesh not initializing allocator states (#19740)
### Ticket
None

### Problem description
On the branch where we have CCLs working,
`test_all_gather_multiple_submeshes` is failing with a segfault. The
problem is that we don't seem to be initializing separate allocator
states for these submeshes.

### What's changed
Migrate the code for initializing separate allocator states for unit
submeshes so that initialization happens across any submesh creation.

https://github.com/tenstorrent/tt-metal/actions/runs/14092235536