Fix bugs with kernel binary caching (#19567)
### Ticket
#18786 

### Problem description
If a termination occurs while a kernel's binaries are being generated,
not all of the binaries may be generated and written to the disk. In the
following run, the program sees that the binary directory exists and so
it thinks that all of the kernel's binaries are on the disk, but it
fails when it tries to read in the missing binaries.

If multiple `DataMovementKernel`s are created with the only difference
between them being the `processor` field, they all have the same hash.
When the program tries to compile these kernels, only the first kernel
is compiled. The remaining kernels aren't compiled because the program
sees that there is already an existing hash and so it thinks that these
kernels have already been compiled.

### What's changed
When generating a kernel's binaries, we write a `.SUCCESS` file to disk
which acts as a marker that all of the binaries were successfully
generated. In following runs, we check for the existence of the
`.SUCCESS` file to ensure that all of a kernel's binaries exist on disk.

When generating the hash for `DataMovementKernel`s, we use the
`processor` field in the hash calculation.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
(https://github.com/tenstorrent/tt-metal/actions/runs/14087971790)
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)
(https://github.com/tenstorrent/tt-metal/actions/runs/14087974906)
- [x] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
(https://github.com/tenstorrent/tt-metal/actions/runs/14041072215)
- [x] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
(https://github.com/tenstorrent/tt-metal/actions/runs/14041257919)
- [ ] **(For models and ops writers)** Full [new models
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
CI passes (if applicable)
- [x] New/Existing tests provide coverage for changes