#16007: Avoiding hammering L1 when waiting for other RISCS in firmware
The firmware on slave riscs needs to poll L1 when waiting to start, and the
firmware on the brisc needs to poll when waiting for the other riscs to finish.

This can slow down other memory operations on the core. Add 5 nops in each loop
to reduce that effect, while not introducing much extra latency.