Add flush bit flag to semaphore inc packets (#19765)
This flush mode is required in order to avoid a possible race condition
where a producer sends a payload packet immediately followed by a
semaphore inc payload. This can be a race if the payload is going to one
core (core A), and the semaphore inc goes to another core (core B) and
is used to notify core B that the payload can be read from core A.

Because there is no guarantee of arrival time and ordering across core A
and core B (once the packets leave the fabric cores, it's up to noc),
then core B could get the semaphore update and request a read from core
A before (and get some data from there) before the original payload
packet is fully committed to core A.

The default is to true (flush outstanding writes) before writing out the
semaphore inc. Users can override to no flush if they know there is no
race.

This option is specified when setting the packet header noc command
information.


Note that this PR will not require changes to existing kernels.

To accomplish this and simplify enabling/disabling/parametrizing the new
code path, some refactoring was done to move some common constants and
types to common headers that can be included by the main program and
helper files (headers).

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/14108137345
- [x] Microbench:
https://github.com/tenstorrent/tt-metal/actions/runs/14108034242
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: Sean Nijjar <snijjar@tenstorrent.com>