[Test Infrastructure] Fixed bug in test_tilize_calculate_untilize_L1  (#68)
### Problem description
Once TRISCs were synced to run at the same time in order to resolve
hangs because unpack was lagging behind pack and math TRISCs,
`test_tilize_calculate_utilize.py` started to fail because it attempted
to fuse two LLK pipeline runs. This means that in the first run we
unpack from two buffers in L1 and after math we pack to two other
buffers in L1. Then in the second run we unpack from these two buffers
that pack wrote to, however `unpacker` started to read from the buffers
before the `packer` completed writing to L1.

This worked previously because the `unpacker` was severely lagging
behind `packer` and `math` (approximately 700 cycles), and this lag
allowed for `packer` to finish writing to L1 before the `unpacker` would
start reading. Since `unpacker` is synced with `packer` we must make
`unpacker` wait for packer to finish before it begins reading from L1.

### What's changed
Implemented semaphore/ flag to control the execution flow required
between different components of the system. In
`tilize_calculate_untilize_L1.cpp` file we use `packer` mailbox as a
semaphore flag for inter-process communication mechanism with the
`unpacker`. The mailbox is a shared memory location that packer
processor writes to, and unpacker processor reads from. The packer will
update the value it stores in its mailbox to signal that it has written
to L1. Meanwhile the unpacker will wait/stall until packer signals that
it has finished before it begins.

The use of `__atomic_load_n` with `__ATOMIC_SEQ_CST` means that the
value is being read atomically with a memory synchronization order,
ensuring data consistency and avoiding race conditions without the need
for complex locking mechanisms.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update