initial impl of fused fabric write + atomic inc (#19877)
# Ticket
https://github.com/tenstorrent/tt-metal/issues/19778

# Note on the extent of the changes
This feature introduced a new packet type. Therefore, to accomodate that
in our testing infrastructure, there was a fair bit of refactoring done
to generalize the testing infrastructure further to support differnet
packet types and provide more configurability for the test itself.

# Problem description
A common use case for fabric clients is to send some number of payloads,
followed by a semaphore increment. This can be made more efficient by
letting a client to both of these with a single packet.

# What's changed
Added a new fused write + atomic in packet. The packet allows the client
to specify is flushing is desired.

When flush is enabled (default behaviour), then the fabric will make
sure the payload is committed to the destination bank before writing the
semaphore inc over the noc.

**AT PRESENT, USING `flush=true` MODE FOR FUSED PACKET IS AROUND HALF
(GIVE OR TAKE) THE THROUGHPUT OF RAW WRITE, SO IT IS SOMETIMES WORSE TO
USE THAN A SEPARATE WRITE + ATOMIC INCC**
... until it can be optimized further :) 

This is due to the current implementation which is inefficient in that
it essentially causes erisc to poll for payload write completion before
doing semaphore inc - even though it could do other work in theory.

Therefore it is recommended, as of today, to only use this packet type
when flush mode is not needed by the user.

## Creating a Fused Write + Atomic Inc Packet
Packet setup is similar to other types of packets. A new noc command
header spec is provided via `NocUnicastAtomicIncFusedCommandHeader`
which has the following constructor arguments:

`NocUnicastAtomicIncFusedCommandHeader(uint64_t noc_address, uint64_t
semaphore_noc_address, uint16_t val, uint16_t wrap, bool flush = true)`

The user must opt-in to no-flush behaviour, for safety reasons.

## When is it safe to use no-flush mode?
Generally speaking, it is only safe to specify `flush=false` when the
client knows that the semaphore consumer and payload consumer are the
same core.

If they are not the same, there is a race condition. For example,
consider the below diagram. A fused packet is received over Ethernet and
is intended for the local device.

The first edge (1) indicates the packet data/payload destination. The
second edge (2) indicates the semaphore inc destination.

If the consumer of the semaphore inc uses the semaphore to understand
when the data is available then there is a race without flushes between.
This is because the paths to each receiver can be different lengths and
have different delays (e.g. congestion). In cases like that, then it is
possible for the semaphore inc to arrive first and for the worker to
intiate a read of the data before it even arrives at the final
destination. If that happens, it is possible for the worker to get
incorrect data.

```
                                     
                                     
                    chip             
        ┌──────────────────────────┐ 
        │ erisc                    │ 
        │┌──────┐       ┌────┐     │ 
        ││      │1)     │    │     │ 
────────┼┼►     ┼──────►│Bank│     │ 
        ││      ┼───┐   │ 0  │     │ 
        ││      │2) │   └────┘     │ 
        ││      │   │              │ 
        ││      │   │   ┌────┐     │ 
        │└──────┘   └──►│ W  │     │ 
        │               │    │     │ 
        │               └────┘     │ 
        └──────────────────────────┘ 

```       

For this reason, the fabric defaults to flush=true because it has no way
of understanding the program semantics.

## Performance 
**For the time being, it is recommended be careful when using fused
write + atomic inc if `flushed=true` because the performance delta is
quite large over raw writes**

Broadly speaking, raw writes are 7-12% faster than no-flush fused packet
write+atomic. This makes it advantageous to use when no-flush is an
option for the given workload. However, when flush mode is enabled, then
raw writes are anywhere from 1.6x to 2x faster.

Some measurements are shown below:

| Type | Packet Size (B) | Flush GB/s | No-Flush GB/s | speedup GB/s (x)
| Payload Only Write GB/s | Speedup Over `flush=true`| Speedup Over
`flush=false` |
|-|-|-|-|-|-|-|-|
|fabric_mcast |	16	| 0.023| 0.035| 1.52 |	0.037	| 1.069|	1.634|
|fabric_mcast|	2048|	2.597|	4.522|	1.741|	4.849|	1.072|	1.867|
|fabric_mcast|	4096|	4.280|	9.077|	2.120|	9.688|	1.067|	2.263|
|fabric_unicast|	16|	0.0276|	0.0415|	1.501|	0.0447|	1.078|	1.619|
|fabric_unicast|	2048|	3.165|	5.318|	1.680|	5.740|	1.079|	1.813|
|fabric_unicast|	4096|	5.621|	10.214|	1.817|	11.451|	1.121|	2.037|




### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/14182501628
- [x] Metal Microbench:
https://github.com/tenstorrent/tt-metal/actions/runs/14203549420
- [x] T3K nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/14182612472/job/39732196230
- [x] T3K unit:
https://github.com/tenstorrent/tt-metal/actions/runs/14186206951
- [x] New/Existing tests provide coverage for changes