[Test Infra] Ensure Tensix Operations Complete Before Reading Results from L1 & Wait for Tensix Completion to Expose Hangs (#82)
### Problem description
Previously, we were reading from L1 before verifying if our Tensix
operations completed. This could have resulted in reading invalid data
from L1, leading to misinterpretation of test failures as functional
issues, when in fact the problem might have been a synchronization issue
or a hang.
### What's changed
To address this, we’ve updated the flow to ensure that TRISC calls are
fully completed before we read the results from L1. Additionally, we
replaced our assertion calls with wait calls that pause execution until
the Tensix operation finishes. This change improves our ability to
distinguish between actual test failures (due to incorrect results) and
potential hangs, as the test will now wait for the operation to complete
instead of failing immediately on an assertion, which previously didn’t
provide visibility into how long the assertion had been false.

### Type of change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update