fix: detect CHIP_ARCH from pytest and simplify dev env (#84)
### Ticket
None

### Problem description
Thus far, we had to execute an external script in order to set
`CHIP_ARCH` variable needed for running the tests.
Also, developer environment setup was requiring a lot of steps and took
long to finish. This has now been moved to Docker.

### What's changed
This PR simplifies the process of running the tests and setting up
developer environments.

### Type of change

- [x] New feature (non-breaking change which adds functionality)