Move prefetcher to BRISC
Currently the prefetcher is running up against the NCRISC code size limit.
BRISC on the same core is idle, so we can move the code there and bump the
optimization level to O2 to more than counteract any performance loss.
Performance is now up to 10% better on some benchmarks, so update the golden
files as well.