#0: Simplify semaphore dispatch logic.
populate_dispatch_data was copying the original data into
program_transfer_data, but we are making a bunch of assumptions about the
elements (e.g. each is the same size) and we're not really taking advantage of
the flexibility of the transfer_info struct. If we move the logic to get data
about semaphores into assemble_device_commands, that simplifies the code by
getting rid of the unnecessary conversion.