Fix Conv2D with dilation and shallow channels.
When using Conv2D with dilation, if the input channel alignment is not a multiple of 32, we get incorrect output. The reason for this issue is that the activations reader in the dilation codepath ignores the act_block_w_extra_align_bytes argument.

In the case of the auto_shard codepath, input channel alignment is deduced automatically. This is why auto_shard is enabled in test_conv_dilation to test the codepath where input channels are smaller than 32.
codepath when input channels is smaller than 32.