Remove race conditions in packer thread. (#97)
Matmul hangs non deterministically on P100 scrappy boards but not
production boards. This PR fixes the issue of packer thread updating the
register containing an address used by the packer before the packer mop
finishes. The change added a wait to the packer in an existing STALLWAIT
and a NOP after a WRCFG (without which another race condition may be
there).