Fix matmul ND hang due to packer race condition (#20415)
### Ticket
[Link to Github Issue

](https://github.com/tenstorrent/tt-metal/issues/18779#issuecomment-2782834310)

### Problem description
Matmul hangs and it points to a race condition where we change a
register while the packer is packing. If we just modify a stallwait
instruction to wait for the packer to finish, without making the other
changes made in the code (for future-proofing), the issue is fixed.
Changing the stallwait should not cause any timing change in the absence
of the bug. The question of whether the address gets latched to the
Packer remains (confirming with hardware), but the fact that modifying
the stallwait (that does not cause any timing change in absence of a
bug) fixes this indicates may be a hardware bug related to improper
latching, but in anycase the solution remains the same.

### What's changed
We wait till the packer is finished before updating the register that
contains an address the packer is supposed to use.
The matmul specific changes are in
[tt_llk_blackhole/common/inc/cpack_common.h](https://github.com/tenstorrent/tt-llk/compare/d25f5d582657f4d76402338f830b05726197afa8...3db62802e5ae60fa01cc4a8d7da273944d99157e#diff-e14aeb60d4d01c1f2ff73a1c81c870e64b40be8335ab49092277e8c895e679c2)

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)
- [ ] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes