#20414: Fix conv2d hang with large kernel
In Conv2dConfg we have an inconsistent default
between cpp code python code (pybind).

Fixing the default on python side unblocked some
test cases as they are not running out of memory
anymore.

Once they started executing, they thit a hang
in conv2d activations reader kernel.
We were allocating an array on stack with size
kernel_h * kernel_w and data type is uint32_t.
In case of large kernel size like 32x32 this is
4k bytes on stack. Which was blowing over our stack.

Code is adjusted not to use this stack allocated
array to avoid this issue.