Copy cluster descriptor to report directory when logging is enabled (#19862)
### Problem description

A copy of the `cluster_descriptor.yaml` file is required to implement
new features in TTNN-Visualizer. It is currently being generated, but it
is getting saved in `/tmp` only. We would like to have a copy of the
`cluster_descriptor.yaml` file saved in the report directory when
logging is enabled.

### What's changed

The `ttnn/ttnn/decorators.py` has been modified to check if there is a
`cluster_descriptor.yaml` file in the report directory. If it already
exists, it does nothing. If it does not exist, it runs a function being
added here that copies the file from `/tmp` to the configured report
directory. The logic in the function looks for the cluster descriptor
file with the most recent created timestamp in `/tmp`, and uses that
one.

My first preference was to try to get the path to the file from
somewhere in the Python code rather than looking for the file with a
shell pattern, but I could not find the path to the file being exposed
in Python anywhere. I traced through the code and I see that it gets
generated in the
[tt_ClusterDescriptor::serialize_to_file()](https://github.com/tenstorrent/tt-umd/blob/main/device/tt_cluster_descriptor.cpp#L1012-L1022)
method in TT-UMD. But I couldn't find where the return value was being
stored anywhere accessible to the TTNN Python code.

An alternative approach to searching for the file with `glob.glob` would
be to expose that C++ function to the Python code, so we can get the
path that way instead. Since it's not exposed already, as far as I can
tell, looking for it with
`glob.glob(f"/tmp/umd_*/cluster_descriptor.yaml")` was an easier way to
do it.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)
- [ ] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes