llama3_subdevices vLLM follow up (#20628)
### Ticket
n/a

### Problem description
We've updated vLLM interface for the SamplingParams and this was
reflected in tt_transformers in
https://github.com/tenstorrent/tt-metal/commit/ff68e3c34184037f051d9a2a5f4227a52bed3765

llama3_subdevices need a follow up.

### What's changed
SamplingParams are used, `is_token` replaces `argmax_on_device`

### Checklist
- ✅ All post commit -
[29859](https://github.com/tenstorrent/tt-metal/actions/runs/14408484494)
- ✅ (Single-card) Fast dispatch frequent tests -
[839](https://github.com/tenstorrent/tt-metal/actions/runs/14408508608)
- ⚠️ (T3K) T3000 unit tests -
[3639](https://github.com/tenstorrent/tt-metal/actions/runs/14408514814),
failed in the same places as on main, unrelated to my changes. It's been
resolved on main since.