[Test Infrastructure] Fixing how data is read from L1 and written to L1 (in bytes) (#110)
### Problem description
There is a reoccurring issue where data written to L1 is not organized
in same order as when packer writes data to L1, in other words our
program was writing incorrectly to L1 and this caused comparison of
golden and result to be incorrect even though the numbers produced were
accurate they were out of order.

Once this issue was addressed and solved, our data was in stored uniform
order throughout, however we were reading words from L1 using tt-exalens
but misinterpreting in which order to read the bytes, and according to
which format. Since this would cause too much logic and work around to
develop our own model for how we read bytes from a word in L1, instead
we now use tt-exalens function `read_from_device` which reads bytes from
words in L1 correctly.

Now there is no need for bug patches made prior when reading from device
in certain formats like bfp8_b or float16, everything is uniform and
bytes are read in consecutive order for each format except the number of
bytes varies according to size of format/datum.

### What's changed
In `device.py` instead of calling `read_words_from_device` and from
there extracting bytes and interpreting them further, instead we now
call function `read_from_device` from exalens which gives us a string of
bytes that we later convert to a list of bytes in `unpack.py` and
convert to wanted (pack_dst) format accordingly.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update