[Test Infra] Added flexibility for dynamic byte count calculation for reading datums of different sizes (#111)
### Problem description
<!-- Provide context for the problem. -->
Instead of reading words from device to extract our result vector, we
instead call read_from_device which reads the data in bytes. Previously
this was set to read a fixed number or words and then a fixed number of
bytes (maximum amount of bytes because it was set to read bytes for
32bit datums).
However, we want to be able to be flexible with our stimuli generation
for testing, which means input tensors will be of different sizes and
different formats, thus the number of bytes we read should be flexible
to these changes, instead of fixed.

### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
Now we read the exact number of bytes necessary, we take into account
the size of our output vector and the size of our datum, defined by
`format_num_bytes` map. This calculation is done in
`utils.py::calculate_read_byte_count` and the result is passed into
exalens' `read_from_device` function called in `device.py`.
### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)