Be agnostic about target triple's manufacturer (#117)
We plan to change the target triple from `riscv32-unknown-elf` to
`riscv32-tt-elf`, to reduce the possibility of confusion between riscv
toolchains. This changes the Makefile so that it doesn't care what the
manufacturer component is.

I notice that a linker script was incorrectly adding
`/opt/riscv32i/riscv32-unknown-elf/lib` as a library search path. It
should not be looking there.