Add core placement opt to 1d fabric (#20787)
Add core placement logic for 6U to avoid congestion issue on multiple
links.

Summary is here:
basically, 
4 link linear mcast improve from 6.33 to 9.33, and full ring from 6.48
to 8.94, some other 3/4 links tests also improved.
the changes are to choose between noc0 and noc1 based on which one is
more optimal to use. local write is always noc1, since this is the
optimal noc. This means when we check flush, we need to check both noc
status, which has some overhead.

the tests less than 4KB got worse in some senerios, perhaps because the
sw overhead of checking two nocs
ring test only works best for 4 links, for 3 link test, it got worse in
some test, perhaps because ring is more sw bound


[bandwidth_summary_core.csv](https://github.com/user-attachments/files/19796005/bandwidth_summary_core.csv)

### Checklist
- [x] [All post commit]
https://github.com/tenstorrent/tt-metal/actions/runs/14501961127
- [x] ubenchmark
https://github.com/tenstorrent/tt-metal/actions/runs/14516601449
- [x] T3k unit
https://github.com/tenstorrent/tt-metal/actions/runs/14502557575
- [ ]