Remove ARCH_NAME cruft from wheel (#20835)
### Problem description
The wheel has legacy code that was generating an .ARCH_NAME file in the
wheel, which is no longer necessary.
ARCH_NAME was also being used to determine the name of the wheel... it
was appending the tag "any" to the wheel file name.

### What's changed
rm -rf

### Checklist
- [x] [ttnn post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/14522207134)
CI passes
- [x] New/Existing tests provide coverage for changes