Add "Full Cluster" Variants of Fabric BW Tests (#20893)
Current fabric BW tests only run on a signle device sequence - which is
sufficient for evaluating fabric router performance. However, it is
insufficient for testing the hardware or identifiying hardware issues.


### What's changed
This change add a basic "full mesh" version of the fabric test. Not that
the diff is quite large but it is mostly refactoring and cleaning up
code to enable this functionality. Note that this adds the capabilities
for this type of testing but only t3k variants of the tests are added.
6u variants with performance tracking will be added when machines are
available.


### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/14553834162
- [x] Metal Microbenchmark:
https://github.com/tenstorrent/tt-metal/actions/runs/14553835354


https://github.com/tenstorrent/tt-metal/issues/20211