Add bitflip test for galaxy (#20940)
### Ticket
- #10673

### Problem description
There is a known issue on some 4U system, where increased DRAM BW can
cause bitflips in the data. This has been impacting the Llama TG
project, as the increase in decode performance has led to more
instability and ND behavior, caused by this bitflipping.

### What's changed
This PR adds in a pytest that can be used as a "smoke" test to run on
machines to check whether they are susceptible to bit flipping. The test
produces a file `nd_failure_summary.txt` that showcases which devices
have experienced a bit flip, and where in the tensor.
- long test: 10k iters (est. 14 min)
- smoke test: 1k iters (est. 2 min)