#20895: Restrict using zero copy chunking to contiguous outermost dim (#21015)
### Ticket
#20895 

### Problem description
The existing code has a problem with data that needs to be along non
contiguous dimensions.

`tt::stl::Span` doesn't provide a way to express this, so at the moment
disabling this until we find a convenient type that will be able to.

### What's changed
Dropped `dim` argument from `chunk` API and updated the tests
accordingly.

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/14602657237)
- [X] New/Existing tests provide coverage for changes