add yml version of support cloud request template (#20954)
### Ticket
#20447 

### Problem description
Create standardized issue templates across all repos. For unique issue
templates such as this one, the template is converted to yml to match
the others.

### What's changed
Describe the approach used to solve the problem.
Converted a .md template for support cloud requests to be .yml. The
outcome is a file that matches the style and presentation of the other
available templates.

### Checklist
Because pre-commit run for all files results in hundreds of files being
changed, pre-commit was run solely on the new yml file using the --files
option.