Update the support request template to be yml (#21139)
### Ticket
#20447 

### Problem description
This pull request is part of the effort to standardize issue templates
so they use yml. It creates text
boxes the user can type in. 

### What's changed
Describe the approach used to solve the problem.
Convert the .md template to be .yml. In this style, it now matches
templates across all the Forge repos. The impact is that it helps create
more consistent presentation when users go to file issues or requests.

### Checklist
- [ ] Run pre-commit.