Start of -dev package (#21035)
### Ticket
#7915

### Problem description
We have no easy way for projects to consume TT-Metalium. TT-NN is lucky
that it builds in the same tree. But customers do not and should not.
Also other external projects like TT-Forge. Current workarounds are
brittle and painful.

### What's changed
The rough-in of a -dev package.

What it does:
* Builds a .deb package for TT-Metalium SDK
* Builds a .deb package for TT-Metalium examples
* Able to build hello_world_compute_kernel in a clean Ubuntu 22.04
container.
What it does not:
* Not yet functional (waiting for
https://github.com/tenstorrent/tt-umd/pull/777 ).
* Only tested for _building_ one hello kernel example; have not
exercised running it, or building others. Will expand in follow-ups.
* Exercised this in CI. Can't do that until the UMD fix reaches
TT-Metalium.