Fix non tile multiple input shard
Conv2d is failing in case when input shard
is not a multiple of tile size. Bug is in
sliding windows infra which is trying to deduce the
shard height. Instead of decuding the shard height
just pass in actual shard height.

Motivation for this change comes from work of
improving device perf on unet shallow model, where
changing the core count of the model exposed this
problem. max_pool2d on higer core count is prodcing
shard height of 671 which isn't multiple of 32 and
subsequent conv2d has pcc issues.