Un-break build-static-libs (#21273)
### Ticket
#21262

### Problem description
We have a --build-static-libs option in our build_metal.sh script. We
shouldn't break it if we're going to keep it.

### What's changed
Disable packaging if the user has asked for static libs.
In the future we can consider making that work (will need to check
whether OBJECT libs give us what we want there). But we've got circular
deps to sort out before we get that far.