[Test Infrastructure] Data format Inference Model  (#143)
### Ticket
<!-- Link to Github Issue -->
https://github.com/tenstorrent/tt-llk/issues/80
### Problem description
Tests support Input and Output format parametrization meaning that the
user can define the format they what their input tensor in L1 to be and
the format of their output in L1. However for the LLK pipeline to run we
need to define the intermediary formats including unpacker formats, fpu
formats and packer formats. This the user does not define and is left up
to us to define according to the best combination which was developed
during the format sweep effort.

### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
The inference model looks at the input and output format defined by
client, and dest accumulation, if it is on or off and sets the unpacker
and packer src and dest formats accordingly. When dest accumulation is
off the packer accepts the format seen in the dst register and then
convert to the output formats wanted. However when dest accumulation is
on then `Float32` is in dst register and packer gasket can covert to the
output format before the data goes into packer for reformatting; thus
the packer accepts the source format to be the wanted output format and
does no conversion.
- Note that when our wanted output if `Float16_a` and dest accumulation
is turned on we cannot have packer_src and packer_dst formats be the
same because the packer gasket cannot perform conversion from `Float32`
to `Float16_a`, so the packer_src format is `Float32` and packer_dst
format is `Float16_a`. In wormhole architecture, the packer formats are
decided specifically according to dest accumulation.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)